feat: Implement root bootstrap capability handoff
- Update `bharat_bootstrap_info_t` to structurally append `namesvc_endpoint` and `service_receive_endpoint`.
- Fix capability table `owner_pid` to use `uint64_t` corresponding to process IDs for actual scope validation.
- Implement `CAP_TYPE_BOOTSTRAP` natively with `CAP_RIGHT_BOOTSTRAP_LAUNCH` and `CAP_RIGHT_BOOTSTRAP_BIND`.
- Enforce transactional rollback correctly within `ipc_endpoint_create()` if receiving capability fails to grant.
- Correctly seed `SELF_PROCESS` capability using a proper `object_ref` to the process pointer and explicitly defer `namesvc` receive capability allocation to the process manager.
- Prove evidence markers via `BH_SYS_CAPABILITY_DELEGATE` from userspace, emitting assertions only upon operational proof.
- Write robust kernel tests verifying rejection of out-of-scope allocations, lack of escalation, and transaction compensations.

+++ core/kernel/src/cap/cap_dispatch.c
@@ -214,7 +214,7 @@ static kstatus_t cap_validate_object_type_internal(cap_type_t entry_type, cap_ty
     return K_OK;
 }
 
-static kstatus_t cap_validate_scope_internal(const capability_table_t *table, uint32_t requester_pid) {
+static kstatus_t cap_validate_scope_internal(const capability_table_t *table, uint64_t requester_pid) {
     if (requester_pid == 0) {
         return K_OK;
     }
